Refresh mVolumes again just before we add volumes to MTP
In MtpService, mVolumes is the list of volumes queried on
Oncreate() i.e.only when mtp is started.If external sdcard was
not present, then mVolumes will not have sdcard entry.
Subsequently, if user inserts sdcard now, sdcard won't be shared
to PC(via Mtp), as sdcard entry is not present in mVolumes.
This gerrit will refresh mVolumes to again fetch volumes, so
that sdcard entry will be present.

Change-Id: Ifb0a57958ea3521b2af1230d43ef9c04c504d305
Signed-off-by: mydongistiny <jaysonedson@gmail.com>